EICNET-XXXX: Split default and dummy content in 2 separate modules.
