+++ lib/modules/eic_default_content/modules/eic_dummy_content/eic_dummy_content.info.yml
@@ -1,9 +1,10 @@
-name: EIC Default Content
+name: EIC Dummy Content
 type: module
-description: Provides default content to use for all environments.
+description: Provides dummy content to use on local/development environments.
 package: European Innovation Council
 core: 8.x
 php: 7.3
 core_version_requirement: ^8 || ^9
 dependencies:
   - data_fixtures:data_fixtures
+  - eic_default_content:eic_default_content
