+++ lib/modules/eic_default_content/modules/eic_dummy_content/eic_dummy_content.services.yml
@@ -1,0 +1,30 @@
+services:
+  eic_default_content.user_generator:
+    class: Drupal\eic_dummy_content\Generator\UserGenerator
+    arguments: [ '@entity_type.manager' ]
+    tags:
+      - { name: data_fixtures, priority: -1 }
+  eic_default_content.term_generator:
+    class: Drupal\eic_dummy_content\Generator\TaxonomyGenerator
+    tags:
+      - { name: data_fixtures, priority: -2 }
+  eic_default_content.group_generator:
+    class: Drupal\eic_dummy_content\Generator\GroupGenerator
+    tags:
+      - { name: data_fixtures }
+  eic_default_content.group_type_event_generator:
+    class: Drupal\eic_dummy_content\Generator\GroupTypeEventGenerator
+    tags:
+      - { name: data_fixtures }
+  eic_default_content.organisation_generator:
+    class: Drupal\eic_dummy_content\Generator\OrganisationGenerator
+    tags:
+      - { name: data_fixtures }
+  eic_default_content.story_generator:
+    class: Drupal\eic_dummy_content\Generator\StoryGenerator
+    tags:
+      - { name: data_fixtures }
+  eic_default_content.news_generator:
+    class: Drupal\eic_dummy_content\Generator\NewsGenerator
+    tags:
+      - { name: data_fixtures }
